requirements.txt
streamlit
google-generativeai
matplotlib
numpy
sympy
PyPDF2
Pillow
opencv-python-headless